Move console related exceptions to Console/Error
