+++ src/Console/Error/MissingShellMethodException.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * ConsoleException class
+ * MissingShellMethodException class
  *
  * PHP 5
  *
@@ -15,12 +15,16 @@
  * @since         3.0.0
  * @license       MIT License (http://www.opensource.org/licenses/mit-license.php)
  */
-namespace Cake\Error;
+namespace Cake\Console\Error;
+
+use Cake\Error\Exception;
 
 /**
- * Exception class for Console libraries.  This exception will be thrown from Console library
- * classes when they encounter an error.
+ * Used when a shell method cannot be found.
  *
  */
-class ConsoleException extends Exception {
+class MissingShellMethodException extends Exception {
+
+	protected $_messageTemplate = "Unknown command %1\$s %2\$s.\nFor usage try `cake %1\$s --help`";
+
 }
